legacy/base_requirements.txt
amazon-braket-default-simulator==1.20.1
amazon-braket-schemas==1.20.2
amazon-braket-sdk==1.70.2
boto3
botocore
cirq==1.3.0
cirq_core==1.3.0
cirq_ionq==1.3.0
matplotlib==3.5.1
numpy==1.24.4
oqc-qcaas-client
pyGSTi==0.9.12.1
qiskit==0.45.2
qiskit_aer==0.13.2
qiskit_braket_provider==0.0.5
qiskit_terra==0.45.2
scipy
tqdm==4.65.0
jupyter
matplotlib==3.5.1
pylatexenc